Add more full-featured neovim configuration
I've started using neovim more often on my machines and it's
configuration is in Lua. This patch moves the init.vim file that I had
originally to init.vim.minimal for when I just want to use the existing
vimrc (like on a server) and then my full init.lua configuration for the
"full fat" neovim experience with parsers, linters, and LSPs.